requirements.txt
# ─── Core ─────────────────────────────────────────────────────────
pyspark==3.5.0
delta-spark==3.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ─── Kafka ────────────────────────────────────────────────────────
kafka-python==2.0.2
confluent-kafka==2.3.0

# ─── API Serving ──────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.25.0
websockets==12.0

# ─── Data Quality ─────────────────────────────────────────────────
great-expectations==0.18.0

# ─── Monitoring ───────────────────────────────────────────────────
prometheus-client==0.19.0

# ─── Orchestration ────────────────────────────────────────────────
apache-airflow==2.8.0

# ─── Testing ──────────────────────────────────────────────────────
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.0

# ─── Utilities ────────────────────────────────────────────────────
python-dotenv==1.0.0
requests==2.31.0
